refactor: Move quiz creation to socket
- move  `createQuiz` controller function and integration of `Classroom` model from `quizController.js` to a new module `socket.quiz.js`.
- refactor `createQuiz` to utilize socket-related `context` features, instead of relying on `http` request/response model in `socket.quiz.js`.
- remove integration of `createQuiz` and its respective API route endpoint in `quizRoutes.js`.
